Cleaning up to have smaller dependency fingerprint
