+++ Project.toml
@@ -4,8 +4,8 @@ authors = ["Isaia Nisoli <nisoli@im.ufrj.br>, Federico Poloni <federico.poloni@u
 version = "0.2.2"
 
 [deps]
-AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
 CRlibm = "96374032-68de-5a5b-8d9e-752f78720389"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
@@ -14,20 +14,17 @@ DualNumbers = "fa6b7ba4-c1ee-5f82-b5fc-ecf0adba8f74"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FastRounding = "fa42c844-2597-5d31-933b-ebd51ab2693f"
 FastTransforms = "057dd010-8810-581a-b7be-e3fc3b93f78c"
-GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
-Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 IntervalOptimisation = "c7c68f13-a4a2-5b9a-b424-07d005f8d9d2"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
@@ -39,19 +36,16 @@ AbstractFFTs = "1"
 Adapt = "4"
 Arpack = "0.5"
 CRlibm = "1"
-CUDA = "5"
+CUDA = "5.5"
 Documenter = "1"
 DualNumbers = "0.6"
 FFTW = "1"
 FastRounding = "0.3"
 FastTransforms = "0.16"
-GLPK = "1"
-Glob = "1"
 IntervalArithmetic = "0.20.9"
 IntervalOptimisation = "0.4"
 LaTeXStrings = "1"
 Plots = "1"
-Polyhedra = "0.7"
 ProgressMeter = "1"
 RecipesBase = "1"
 Revise = "3"
@@ -61,10 +55,28 @@ SymbolicUtils = "3"
 Symbolics = "6"
 TaylorModels = "0.7"
 TaylorSeries = "0.17"
-julia = "1"
+julia = "1.9"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test"]
+
+[weakdeps]
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
+SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+TaylorModels = "314ce334-5f6e-57ae-acf6-00b6e903104a"
+
+
+[extension]
+CUDAExt = ["CUDA", "Adapt"]
+PlotsExt = ["Plots", "RecipesBase"] 
+SymbolicExt = ["Symbolics", "SymbolicUtils"]
+TaylorModelsExt = "TaylorModels"
+
