+++ ext/CUDAExt/CUDAExt.jl
@@ -1,0 +1,7 @@
+module CUDAExt
+
+using RigorousInvariantMeasures, CUDA
+
+include("NoiseKernel2.jl")
+
+end
