+++ ext/SymbolicsExt/SymbolicsExt.jl
@@ -1,0 +1,4 @@
+module SymbolicsExt
+    using RigorousInvariantMeasures, Symbolics
+    include("HigherDFLY.jl")
+end
