+++ src/InducedLSV.jl
@@ -264,10 +264,7 @@ function iterate_LSV(x, i, α)
     return x
 end
 
-using RecipesBase
-#COV_EXCL_START
-@recipe f(::Type{ApproxInducedLSV}, D::ApproxInducedLSV) = x -> plottable(D, x)
-#COV_EXCL_STOP
+
 
 ChebOrHatNP = Union{Chebyshev,HatNP}
