+++ test/TestInterface/TestInterfaceIndex.jl
@@ -3,5 +3,5 @@ include("TestAssemble.jl")
 include("TestEstimate.jl")
 include("TestNormOfPowers.jl")
 include("TestNormOfPowersNoise.jl")
-include("TestObservables.jl")
+# include("TestObservables.jl")
 include("TestPreimages.jl")
